+++ great_international/models/great_international.py
@@ -1233,6 +1233,54 @@ class AboutDitLandingPage(panels.AboutDitLandingPagePanels, BaseInternationalPag
         related_name='+'
     )
 
+    intro = models.TextField(max_length=255, blank=True)
+    section_one_content = MarkdownField(blank=True)
+    section_one_image = models.ForeignKey(
+        'wagtailimages.Image',
+        null=True,
+        on_delete=models.SET_NULL,
+        related_name='+',
+        blank=True
+    )
+
+    how_dit_help_title = models.CharField(max_length=255, blank=True)
+
+    related_page_one = models.ForeignKey(
+        'great_international.AboutDitServicesPage',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+',
+    )
+
+    related_page_two = models.ForeignKey(
+        'great_international.AboutDitServicesPage',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+',
+    )
+
+    related_page_three = models.ForeignKey(
+        'great_international.AboutDitServicesPage',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+',
+    )
+
+    case_study_image = models.ForeignKey(
+        'wagtailimages.Image',
+        null=True,
+        on_delete=models.SET_NULL,
+        related_name='+',
+        blank=True
+    )
+    case_study_title = models.CharField(max_length=255, blank=True)
+    case_study_text = models.TextField(max_length=255, blank=True)
+    case_study_cta_text = models.CharField(max_length=255, blank=True)
+    case_study_cta_link = models.CharField(max_length=255, blank=True)
+
 
 class AboutDitServiceField(panels.AboutDitServiceFieldPanels, models.Model):
     icon = models.ForeignKey(