+++ great_international/models/invest.py
@@ -33,9 +33,13 @@ class InvestInternationalHomePage(
         related_name='+'
     )
 
-    benefits_section_title = models.CharField(max_length=255)
+    teaser = models.TextField(blank=True)
+
+    benefits_section_title = models.CharField(max_length=255, blank=True)
     benefits_section_intro = models.TextField(max_length=255, blank=True)
     benefits_section_content = MarkdownField(blank=True)
+    benefits_section_cta_text = models.CharField(max_length=255, blank=True)
+    benefits_section_cta_url = models.CharField(max_length=255, blank=True)
     benefits_section_img = models.ForeignKey(
         'wagtailimages.Image',
         blank=True,
@@ -46,7 +50,7 @@ class InvestInternationalHomePage(
     )
 
     capital_invest_section_title = models.CharField(
-        max_length=255
+        max_length=255, blank=True
     )
     capital_invest_section_content = MarkdownField(
         blank=True
@@ -91,44 +95,29 @@ class InvestInternationalHomePage(
         verbose_name="EU exit section image"
     )
 
-    # subsections
-    subsection_title_one = models.CharField(max_length=255, blank=True)
-    subsection_content_one = MarkdownField(blank=True)
-
-    subsection_title_two = models.CharField(max_length=255, blank=True)
-    subsection_content_two = MarkdownField(blank=True)
-
-    subsection_title_three = models.CharField(max_length=255, blank=True)
-    subsection_content_three = MarkdownField(blank=True)
-
-    subsection_title_four = models.CharField(max_length=255, blank=True)
-    subsection_content_four = MarkdownField(blank=True)
-
-    subsection_title_five = models.CharField(max_length=255, blank=True)
-    subsection_content_five = MarkdownField(blank=True)
-
-    subsection_title_six = models.CharField(max_length=255, blank=True)
-    subsection_content_six = MarkdownField(blank=True)
-
-    subsection_title_seven = models.CharField(max_length=255, blank=True)
-    subsection_content_seven = MarkdownField(blank=True)
-
     sector_title = models.TextField(
         default="Discover UK Industries",
-        max_length=255)
+        max_length=255,
+        blank=True
+    )
 
     sector_button_text = models.TextField(
         default="See more industries",
-        max_length=255)
+        max_length=255,
+        blank=True
+    )
 
     sector_button_url = models.CharField(
-        max_length=255)
+        max_length=255,
+        blank=True
+    )
 
     sector_intro = models.TextField(max_length=255, blank=True)
 
     hpo_title = models.CharField(
         max_length=255,
-        verbose_name="High potential opportunity section title"
+        verbose_name="High potential opportunity section title",
+        blank=True
     )
     hpo_intro = models.TextField(
         max_length=255,
@@ -138,7 +127,8 @@ class InvestInternationalHomePage(
 
     setup_guide_title = models.CharField(
         default='Set up an overseas business in the UK',
-        max_length=255)
+        max_length=255,
+        blank=True)
 
     setup_guide_lead_in = models.TextField(
         blank=True,
@@ -153,7 +143,10 @@ class InvestInternationalHomePage(
         related_name='+',
         verbose_name="Setup guide image"
     )
-    setup_guide_call_to_action_url = models.CharField(max_length=255)
+    setup_guide_call_to_action_url = models.CharField(
+        max_length=255,
+        blank=True
+    )
 
     isd_section_image = models.ForeignKey(
         'wagtailimages.Image',
@@ -174,10 +167,55 @@ class InvestInternationalHomePage(
         verbose_name='Investment Support Directory section text'
     )
 
-    how_we_help_title = models.CharField(default='How we help', max_length=255)
-    how_we_help_lead_in = models.TextField(blank=True, null=True)
+    featured_card_one_image = models.ForeignKey(
+        'wagtailimages.Image',
+        blank=True,
+        null=True,
+        on_delete=models.SET_NULL,
+        related_name='+',
+    )
+    featured_card_one_title = models.CharField(blank=True, max_length=255)
+    featured_card_one_summary = MarkdownField(blank=True)
+    featured_card_one_cta_link = models.CharField(max_length=255, blank=True)
+
+    featured_card_two_image = models.ForeignKey(
+        'wagtailimages.Image',
+        blank=True,
+        null=True,
+        on_delete=models.SET_NULL,
+        related_name='+',
+    )
+    featured_card_two_title = models.CharField(
+        max_length=255,
+        blank=True,
+    )
+    featured_card_two_summary = MarkdownField(
+        max_length=255,
+        blank=True,
+    )
+    featured_card_two_cta_link = models.CharField(max_length=255, blank=True)
+
+    featured_card_three_image = models.ForeignKey(
+        'wagtailimages.Image',
+        blank=True,
+        null=True,
+        on_delete=models.SET_NULL,
+        related_name='+'
+    )
+    featured_card_three_title = models.CharField(
+        max_length=255, blank=True
+    )
+    featured_card_three_summary = MarkdownField(blank=True)
+    featured_card_three_cta_link = models.CharField(max_length=255, blank=True)
+
     # how we help
-    how_we_help_text_one = models.CharField(max_length=255)
+    how_we_help_title = models.CharField(
+        default='How we help',
+        max_length=255,
+        blank=True
+    )
+    how_we_help_lead_in = models.TextField(blank=True, null=True)
+    how_we_help_text_one = models.CharField(max_length=255, blank=True)
     how_we_help_icon_one = models.ForeignKey(
         'wagtailimages.Image',
         null=True,
@@ -185,7 +223,7 @@ class InvestInternationalHomePage(
         on_delete=models.SET_NULL,
         related_name='+'
     )
-    how_we_help_text_two = models.CharField(max_length=255)
+    how_we_help_text_two = models.CharField(max_length=255, blank=True)
     how_we_help_icon_two = models.ForeignKey(
         'wagtailimages.Image',
         null=True,
@@ -193,7 +231,7 @@ class InvestInternationalHomePage(
         on_delete=models.SET_NULL,
         related_name='+'
     )
-    how_we_help_text_three = models.CharField(max_length=255)
+    how_we_help_text_three = models.CharField(max_length=255, blank=True)
     how_we_help_icon_three = models.ForeignKey(
         'wagtailimages.Image',
         null=True,
@@ -201,7 +239,7 @@ class InvestInternationalHomePage(
         on_delete=models.SET_NULL,
         related_name='+'
     )
-    how_we_help_text_four = models.CharField(max_length=255)
+    how_we_help_text_four = models.CharField(max_length=255, blank=True)
     how_we_help_icon_four = models.ForeignKey(
         'wagtailimages.Image',
         null=True,
@@ -209,7 +247,7 @@ class InvestInternationalHomePage(
         on_delete=models.SET_NULL,
         related_name='+'
     )
-    how_we_help_text_five = models.CharField(max_length=255)
+    how_we_help_text_five = models.CharField(max_length=255, blank=True)
     how_we_help_icon_five = models.ForeignKey(
         'wagtailimages.Image',
         null=True,
@@ -220,16 +258,22 @@ class InvestInternationalHomePage(
     how_we_help_text_six = models.CharField(max_length=255, blank=True)
     how_we_help_icon_six = models.ForeignKey(
         'wagtailimages.Image',
-        null=True,
         blank=True,
+        null=True,
         on_delete=models.SET_NULL,
-        related_name='+'
+        related_name='+',
     )
 
-    contact_section_title = models.CharField(max_length=255)
+    contact_section_title = models.CharField(max_length=255, blank=True)
     contact_section_content = models.TextField(max_length=255, blank=True)
-    contact_section_call_to_action_text = models.CharField(max_length=255)
-    contact_section_call_to_action_url = models.CharField(max_length=255)
+    contact_section_call_to_action_text = models.CharField(
+        max_length=255,
+        blank=True
+    )
+    contact_section_call_to_action_url = models.CharField(
+        max_length=255,
+        blank=True
+    )
 
 
 class InvestHighPotentialOpportunitiesPage(